exfat_{core,super}.c: fix build on 5.0-rc1, MS_* -> SB_*
In torvalds/linux@e462ec50cb a new set of superblock flags was added
to replace the original MS_* ones, and in torvalds/linux@e262e32d6b
the MS_* flags are suppressed unless uapi/linux/mount.h is included.
As is suggested, we should use the new API now.